// src/app/constants/survey-constants.enum.ts
export enum SurveyConstants {
    COMPLETED_COMPLETION_STATUS = "SURVEYSTATUS.COMPLETE",
    SUBMITTED_COMPLETION_STATUS = "SURVEYSTATUS.SUBMITTED",
    PARTIAL_COMPLETION_STATUS = "SURVEYSTATUS.PARTIAL",
    NOTSTARTED_COMPLETION_STATUS = "SURVEYSTATUS.NOTSTARTED",
    OPEN_TIMELINE_STATUS = "SURVEYTIMELINESTATUS.OPEN",
    DUE_FUTURE_TIMELINE_STATUS = "SURVEYTIMELINESTATUS.DUEFUTURE",
    PAST_DUE_TIMELINE_STATUS = "SURVEYTIMELINESTATUS.PASTDUE",
    SURVEY_MEDIUM_MOBILE = "SURVEYMEDIUM.MOBILE",
    SURVEY_MEDIUM_WEB = "SURVEYMEDIUM.WEB",
    REQUIRES_SIGNATURE = "REQUIRESSIGNATURE",
    PARTIAL_NOT_SIGNED_COMPLETION_STATUS = "SURVEYSTATUS.PARTIALNOTSIGNED",
    ONHOLD_OVERRIDE_STATUS = "SURVEYOVERRIDESTATUS.ONHOLD",
    ENDED_OVERRIDE_STATUS = "SURVEYOVERRIDESTATUS.ENDED",
    FORCED_OPEN_OVERRIDE_STATUS = "SURVEYOVERRIDESTATUS.FORCEDOPEN"
}